remove cache objects usage from todebug and use a model instead.
cache objects takes a couple minutes to run on my database and this
code will block the main thread until it finishes, then clears it and
calls for a refresh. new code simply runs a sql query to get at just
the objects we want using toeventquery and is much faster (and doesn't
block).