Reformatted source code
Right now there is lot of various programming styles used everywhere in source
code files. That makes the source code look like a mess and it also make it hard
to read. I created an astyle rule from most frequently used style in most files
(which is actually some that resembles the Microsoft Visual Studio style) and
applied them to all other files that use different style, so that everything is
consistent now.